requirements.txt
numpy>=1.15.4
opencv-python>=3.4.4.19
pafy>=0.5.4
Pillow>=5.4.1
requests>=2.21.0
termcolor>=1.1.0
urllib3>=1.24.1
youtube-dl>=2019.3.18
colorama>=0.4.1